hardcore ois and arch for build

+++ build.sh
@@ -10,10 +10,12 @@ echo "Building trsl8 $VERSION  os:$OS arch:$ARCH"
 
 cd docker
 docker build --tag gonzague/trsl8-build:latest --tag gonzague/trsl8-build:${VERSION} -f Dockerfile_build .
+if [ $? -ne 0 ]; then   exit 1; fi
 cd -
 
 #Compressing
 docker run -v "${PWD}:/go/src/github.com/lamarios/trsl8" gonzague/trsl8-build
+if [ $? -ne 0 ]; then   exit 1; fi
 #tar -czf trsl8-$VERSION-$OS-$ARCH.tar.gz trsl8
 
 echo "Building the Web"
@@ -22,7 +24,9 @@ rm -Rf static/*
 
 cd web/
 npm install
+if [ $? -ne 0 ]; then   exit 1; fi
 npm run build
+if [ $? -ne 0 ]; then   exit 1; fi
 cd -
 
 rm -Rf docker/static
@@ -32,4 +36,5 @@ cp -R static docker/static
 
 cd docker
 docker build --tag gonzague/trsl8:latest --tag gonzague/trsl8:${VERSION} .
+if [ $? -ne 0 ]; then   exit 1; fi
 cd -
